--- Servidor/src/controllers/index.js
@@ -1,1 +1,0 @@
-// controllers
